requirements.txt
attrs==23.1.0
cachetools==5.3.1
Cartopy==0.21.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.0
cycler==0.11.0
earthengine-api==0.1.357
et-xmlfile==1.1.0
Fiona==1.9.4.post1
fonttools==4.40.0
geopandas==0.13.2
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyodbc==4.0.39
pyparsing==3.1.0
pyproj==3.5.0
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
shapely==2.0.1
six==1.16.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
xarray==2023.1.0
zipp==3.15.0
